+++ org.eclipse.paho.client.mqttv3/src/main/java/org/eclipse/paho/client/mqttv3/internal/wire/MultiByteInteger.java
@@ -20,14 +20,14 @@ package org.eclipse.paho.client.mqttv3.internal.wire;
  * specification.
  */
 public class MultiByteInteger {
-	private long value;
+	private int value;
 	private int length;
 	
-	public MultiByteInteger(long value) {
+	public MultiByteInteger(int value) {
 		this(value, -1);
 	}
 	
-	public MultiByteInteger(long value, int length) {
+	public MultiByteInteger(int value, int length) {
 		this.value = value;
 		this.length = length;
 	}
@@ -42,7 +42,7 @@ public class MultiByteInteger {
 	/**
 	 * @return the value of this MBI.
 	 */
-	public long getValue() {
+	public int getValue() {
 		return value;
 	}
 }
